% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sprite.R
\name{theme_sprite}
\alias{theme_sprite}
\title{sprite_theme}
\usage{
theme_sprite()
}
\value{
panel.background = element_rect(fill = "#02A04C")

panel.grid.major.x = element_line(colour = "#F7D704", linetype = 3, size = 0.5)

panel.grid.minor.x = element_blank()

panel.grid.major.y =  element_line(colour = "#FFFFFF", linetype = 3, size = 0.5)

axis.text = element_text(colour = "#FFFFFF")

axis.title = element_text(colour = "#1e5cb3")
}
\description{
exactly what is sounds like
}
\details{
see documentation for...

theme(),
element_rect(),
element_line(),
element_blank(),
element_text()
}
\examples{
add to ggplot as u would normally add themes
ggplot(data, aes(x, y)) + theme_sprite()
}
